Move catch.hpp from tests/include to lib/catch2
tests/ contains autests, not unit tests, so the catch header
shouldn't be isolated there. It is easier to locate in lib/
and will allow all potential CMake targets for bundled third-party
software to be in one place.
