+++ plugins/prefetch/Makefile.inc
@@ -29,5 +29,5 @@ prefetch_prefetch_la_SOURCES = \
 
 check_PROGRAMS += prefetch/test_evaluate
 
-prefetch_test_evaluate_CPPFLAGS = $(AM_CPPFLAGS) -I$(abs_top_srcdir)/tests/include -DPREFETCH_UNIT_TEST
+prefetch_test_evaluate_CPPFLAGS = $(AM_CPPFLAGS) -I$(abs_top_srcdir)/lib/catch2 -DPREFETCH_UNIT_TEST
 prefetch_test_evaluate_SOURCES = prefetch/test/test_evaluate.cc prefetch/evaluate.cc prefetch/common.cc
