+++ plugins/s3_auth/Makefile.inc
@@ -19,7 +19,7 @@ s3_auth_s3_auth_la_SOURCES = s3_auth/s3_auth.cc s3_auth/aws_auth_v4.cc
 
 check_PROGRAMS +=  s3_auth/test_s3auth
 
-s3_auth_test_s3auth_CPPFLAGS = $(AM_CPPFLAGS) -I$(abs_top_srcdir)/tests/include -DAWS_AUTH_V4_UNIT_TEST
+s3_auth_test_s3auth_CPPFLAGS = $(AM_CPPFLAGS) -I$(abs_top_srcdir)/lib/catch2 -DAWS_AUTH_V4_UNIT_TEST
 s3_auth_test_s3auth_LDADD = $(OPENSSL_LIBS)
 s3_auth_test_s3auth_SOURCES = \
     s3_auth/unit_tests/test_aws_auth_v4.cc \
